requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6

# Document processing
pypdf2==3.0.1
sentence-transformers==2.2.2

# Vector store
faiss-cpu==1.7.4

# AI/ML
google-generativeai==0.3.2

# Discord (optional adapter)
discord.py==2.3.2

# Telegram (future adapter)
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
rich==13.7.0

# Development
pytest==7.4.3
black==23.12.0